+++ bote/__init__.py
@@ -5,5 +5,5 @@ from bote.__main__ import Mailer
 from bote import _version
 
 NAME = "bote"
-__version__ = f"{_version.__version__}"
+__version__ = _version.__version__
 __author__ = "Rüdiger Voigt"